Add resources for lib pr (#85)
Hello! In preparation to merge https://github.com/DataDog/documentation/pull/10532 I've taken the liberty of stubbing out the code & test to add support for `Datadog::Users::User` (#27) and `Datadog::Monitors::Downtime` (#26).

Caveats: I'm relatively new to JavaScript, and I've never touched TypeScript (or Jest) before, so it's likely that I've made some mistakes. Hopefully this is enough to get you over the finish line. :) Happy to work with you further on this if you like.